Server. Контроллер задач
